feat: adopt bon builders with buildless child composition
