+++ tellur-core/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bon = "3.9.1"
 bytes = "1.11.1"
 fontconfig = "0.10"
 fontdb = "0.23"
