+++ tellur-core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod builder;
 pub mod color;
 pub mod composite;
 pub mod dyn_compare;
@@ -20,4 +21,11 @@ pub mod vector;
 // must be reachable inside this crate too if someone uses them internally.
 extern crate self as tellur_core;
 
-pub use tellur_macros::{raster_component, vector_component};
+pub use tellur_macros::{component, raster_component, vector_component};
+
+// Re-export `bon` so downstream crates and the component macro can reach its
+// runtime (and the `Builder` derive) through `tellur_core` without depending
+// on `bon` directly. The component macro emits `::tellur_core::__bon` paths and
+// sets `#[builder(crate = ::tellur_core::__bon)]` so generated code resolves.
+#[doc(hidden)]
+pub use bon as __bon;
