+++ tellur-core/src/shapes.rs
@@ -12,10 +12,13 @@ use crate::dyn_compare::hash_f32;
 use crate::geometry::{Constraints, Rect, Transform, Vec2};
 use crate::vector::{Fill, Node, Path, PathCommand, Stroke, VectorComponent, VectorGraphic};
 
+#[crate::component(vector)]
 #[derive(Debug, Clone, PartialEq, Hash)]
 pub struct Rectangle {
     pub size: Vec2,
+    #[builder(into)]
     pub fill: Option<Fill>,
+    #[builder(into)]
     pub stroke: Option<Stroke>,
 }
 
@@ -48,10 +51,13 @@ impl VectorComponent for Rectangle {
     }
 }
 
+#[crate::component(vector)]
 #[derive(Debug, Clone, PartialEq)]
 pub struct Circle {
     pub radius: f32,
+    #[builder(into)]
     pub fill: Option<Fill>,
+    #[builder(into)]
     pub stroke: Option<Stroke>,
 }
 
@@ -77,10 +83,13 @@ impl VectorComponent for Circle {
     }
 }
 
+#[crate::component(vector)]
 #[derive(Debug, Clone, PartialEq, Hash)]
 pub struct Ellipse {
     pub radii: Vec2,
+    #[builder(into)]
     pub fill: Option<Fill>,
+    #[builder(into)]
     pub stroke: Option<Stroke>,
 }
 
@@ -147,5 +156,54 @@ fn ellipse_to_graphic(radii: Vec2, fill: Option<Fill>, stroke: Option<Stroke>) -
             stroke,
             transform: Transform::IDENTITY,
         }),
+    }
+}
+
+#[cfg(test)]
+mod builder_tests {
+    use super::*;
+    use crate::builder::VectorBuilderPlacement;
+    use crate::color::Color;
+    use crate::geometry::Anchor;
+    use crate::vector::Paint;
+
+    fn paint() -> Paint {
+        Paint::Solid(Color::rgb_u8(1, 2, 3))
+    }
+
+    #[test]
+    fn complete_builder_converts_into_self_and_box() {
+        // bon derive(Into): a complete builder converts into the struct itself
+        // (so `impl Into<Ellipse>` args accept a builder, no `.build()`).
+        let e: Ellipse = Ellipse::builder().radii(Vec2(3.0, 3.0)).fill(paint()).into();
+        assert_eq!(e.radii, Vec2(3.0, 3.0));
+        assert!(e.fill.is_some());
+
+        // ours: complete builder -> Box<dyn VectorComponent>, no `.build()`.
+        let _boxed: Box<dyn VectorComponent> = Ellipse::builder().radii(Vec2(3.0, 3.0)).into();
+        // ours: a built value -> Box<dyn VectorComponent>.
+        let _boxed2: Box<dyn VectorComponent> = Ellipse {
+            radii: Vec2(1.0, 1.0),
+            fill: None,
+            stroke: None,
+        }
+        .into();
+    }
+
+    #[test]
+    fn builder_place_at_and_anchored_snap() {
+        // place_at on a builder, no `.build()`.
+        let p = Ellipse::builder()
+            .radii(Vec2(5.0, 5.0))
+            .place_at(Vec2(2.0, 3.0));
+        assert_eq!(p.position, Vec2(2.0, 3.0));
+
+        // anchored().snap_to() on a builder: CENTER of a 10x10 box snapped to
+        // (10,10) lands its origin at (5,5).
+        let p2 = Ellipse::builder()
+            .radii(Vec2(5.0, 5.0))
+            .anchored(Anchor::CENTER)
+            .snap_to(Vec2(10.0, 10.0));
+        assert_eq!(p2.position, Vec2(5.0, 5.0));
     }
 }
