+++ tellur-renderer/examples/raster_layer_to_png.rs
@@ -7,6 +7,7 @@
 use std::fs::File;
 
 use tellur_core::color::Color;
+use tellur_core::component;
 use tellur_core::geometry::{Anchor, Vec2};
 use tellur_core::layer::Layer;
 use tellur_core::placement::RasterPlacement;
@@ -14,55 +15,54 @@ use tellur_core::raster::{RasterComponent, Resolution};
 use tellur_core::render_context::PassThrough;
 use tellur_core::shapes::{Circle, Rectangle};
 use tellur_core::vector::Paint;
-use tellur_core::vector_component;
-use tellur_renderer::Rasterizable;
+use tellur_renderer::RasterizableBuilder;
 
 /// A translucent colored circle. The whole shape is parameterised by hue
 /// and radius; saturation, lightness and alpha are baked in.
-#[vector_component]
+#[component(vector)]
 fn blob(radius: f32, hue: f32) -> impl VectorComponent {
-    Circle {
-        radius,
-        fill: Paint::Solid(Color::hsla(hue, 0.7, 0.55, 0.65)).into(),
-        stroke: None,
-    }
+    Circle::builder()
+        .radius(radius)
+        .fill(Paint::Solid(Color::hsla(hue, 0.7, 0.55, 0.65)))
+        .build()
 }
 
 fn main() {
     let scene_size = Vec2(1280.0, 720.0);
-    let scene = Layer {
-        size: Some(scene_size),
-        children: vec![
-            Rectangle {
-                size: scene_size,
-                fill: Paint::Solid(Color::rgb_u8(245, 240, 230)).into(),
-                stroke: None,
-            }
-            .rasterize()
-            .at(Vec2::ZERO),
-            Blob {
-                radius: 200.0,
-                hue: 0.0,
-            }
-            .rasterize()
-            .anchored(Anchor::CENTER)
-            .snap_to(Anchor::new(0.4, 0.4).point(scene_size)),
-            Blob {
-                radius: 200.0,
-                hue: 120.0,
-            }
-            .rasterize()
-            .anchored(Anchor::CENTER)
-            .snap_to(Anchor::new(0.6, 0.4).point(scene_size)),
-            Blob {
-                radius: 200.0,
-                hue: 240.0,
-            }
-            .rasterize()
-            .anchored(Anchor::CENTER)
-            .snap_to(Anchor::new(0.5, 0.65).point(scene_size)),
-        ],
-    };
+    let scene = Layer::builder()
+        .size(scene_size)
+        .child(
+            Rectangle::builder()
+                .size(scene_size)
+                .fill(Paint::Solid(Color::rgb_u8(245, 240, 230)))
+                .rasterize()
+                .place_at(Vec2::ZERO),
+        )
+        .child(
+            Blob::builder()
+                .radius(200.0)
+                .hue(0.0)
+                .rasterize()
+                .anchored(Anchor::CENTER)
+                .snap_to(Anchor::new(0.4, 0.4).point(scene_size)),
+        )
+        .child(
+            Blob::builder()
+                .radius(200.0)
+                .hue(120.0)
+                .rasterize()
+                .anchored(Anchor::CENTER)
+                .snap_to(Anchor::new(0.6, 0.4).point(scene_size)),
+        )
+        .child(
+            Blob::builder()
+                .radius(200.0)
+                .hue(240.0)
+                .rasterize()
+                .anchored(Anchor::CENTER)
+                .snap_to(Anchor::new(0.5, 0.65).point(scene_size)),
+        )
+        .build();
 
     let image = scene.render(scene_size, Resolution::new(1280, 720), &mut PassThrough);
