+++ tellur-renderer/examples/text_to_png.rs
@@ -4,47 +4,43 @@
 
 use std::fs::File;
 
+use tellur_core::builder::VectorBuilderPlacement;
 use tellur_core::color::Color;
 use tellur_core::geometry::{Anchor, Vec2};
 use tellur_core::layer::VectorLayer;
-use tellur_core::placement::VectorPlacement;
 use tellur_core::raster::{RasterComponent, Resolution};
 use tellur_core::render_context::PassThrough;
 use tellur_core::shapes::Rectangle;
-use tellur_core::text::{Text, TextSpan, Weight, SANS_SERIF};
+use tellur_core::text::{Text, TextSpan, SANS_SERIF};
 use tellur_core::vector::Paint;
 use tellur_renderer::Rasterizable;
 
 fn main() {
     let scene_size = Vec2(1920.0, 1080.0);
-    let scene = VectorLayer {
-        size: Some(scene_size),
-        children: vec![
-            Rectangle {
-                size: scene_size,
-                fill: Paint::Solid(Color::rgb_u8(255, 255, 255)).into(),
-                stroke: None,
-            }
-            .at(Vec2::ZERO),
-            Text {
-                font: SANS_SERIF.clone(),
-                size: 96.0,
-                weight: Weight::NORMAL,
-                fill: Paint::Solid(Color::rgb_u8(30, 30, 30)),
-                spans: vec![
-                    TextSpan::plain("Hello "),
-                    TextSpan {
-                        text: "world".into(),
-                        fill: Some(Paint::Solid(Color::rgb_u8(220, 60, 60))),
-                        ..TextSpan::default()
-                    },
-                    TextSpan::plain("!"),
-                ],
-            }
-            .anchored(Anchor::CENTER)
-            .snap_to(Anchor::CENTER.point(scene_size)),
-        ],
-    };
+    let scene = VectorLayer::builder()
+        .size(scene_size)
+        .child(
+            Rectangle::builder()
+                .size(scene_size)
+                .fill(Paint::Solid(Color::rgb_u8(255, 255, 255)))
+                .place_at(Vec2::ZERO),
+        )
+        .child(
+            Text::builder()
+                .font(SANS_SERIF.clone())
+                .size(96.0)
+                .fill(Paint::Solid(Color::rgb_u8(30, 30, 30)))
+                .span("Hello ")
+                .span(
+                    TextSpan::builder()
+                        .text("world")
+                        .fill(Paint::Solid(Color::rgb_u8(220, 60, 60))),
+                )
+                .span("!")
+                .anchored(Anchor::CENTER)
+                .snap_to(Anchor::CENTER.point(scene_size)),
+        )
+        .build();
 
     let image = scene
         .rasterize()
