+++ tellur-renderer/src/lib.rs
@@ -6,7 +6,7 @@ pub mod shadow;
 pub mod video;
 
 pub use outline::Outline;
-pub use rasterize::{Rasterizable, Rasterize};
+pub use rasterize::{Rasterizable, RasterizableBuilder, Rasterize};
 pub use render_context::CachingRenderContext;
 pub use shadow::DropShadow;
 pub use video::{FfmpegEncoder, FfmpegError};
